+++ spec/support/factories.rb
@@ -1,16 +1,27 @@
 module TestHelpers
   module Factories
-    def ingestion_event(attrs = {})
-      defaults = {
-        event_code: 'some-code',
-        event_date: Time.now.round.iso8601(3),
-        transaction_type: 'DeviceBinding'
-        # payload: {
-        #   event_group_id: 'some-transaction-id',
-        #   test_field: 'test value'
-        # }
-      }
-      ::Ubersicht::Ingestion::Event.new(defaults.merge(attrs))
+    def attributes_for(factory_name, attrs = {}) # rubocop:disable Metrics/MethodLength
+      case factory_name
+      when :ingestion_event
+        defaults = {
+          id: 'some-id',
+          type: 'some-code',
+          time: Time.now.round.iso8601(3),
+          source: 'account_1.dauth.device_binding_events',
+          data: {
+            #   event_group_id: 'some-transaction-id',
+            #   test_field: 'test value'
+          }
+        }
+        defaults.merge(attrs)
+      end
+    end
+
+    def build(factory_name, attrs = {})
+      case factory_name
+      when :ingestion_event
+        ::Ubersicht::Ingestion::Event.new(attributes_for(factory_name, attrs))
+      end
     end
   end
 end